Ignore comments in get_symbol()
The get_symbol() method is supposed to skip comments altogether and return to
the parser the symbol that comes right after any comment or a series of lines of
comments. The only case that the parser should know about a comment is when the
comment operator "//" is wrong, i.e. it is a single slash "/". In this case,
the scanner should return a symbol corresponding to the "/", of type
INVALID_SYMBOL, so that the parser can notify the user about the syntax error.

This patch makes sure that the correct comment operator is used (double slash),
and then uses recursion to return the symbol right after the series of comments.