Update error management to work with all errors
Modified error() in Parser to call a funciton to display the error appropriately by displaying the line and type of error with an error message. Now it calls a function to do the display and another for error recovery. Also made it more versatile so it is easier to add semantic errors later on.