Add skeleton of function make_connection
Make_connection is now called when connection definition finishes
and has not enocuntered any error. It calls to helper functions
get_input and get_device that will give the ids to make the correct
connection. This is only a skeleton and has to be improved.

TO DO:
*Build the get_input and get_device functions
*Update make_connection to deal with the following scenarios:
    -One to one
    -Many to one
    -One to Many
    -Many to Many

A good idea might be to store the inputs and outputs in a list.
Storing a series of symbols:
I/O = (identifier, port)
*port can be None a number if an input or CLK, DATA, SET, CLEAR,
Q or QBAR for DTYPE
Inputs = [I1, I2, ...] as specificed above
Similarly for outputs