Change signal() to return (name, port)
This is useful because now signal can be called by connection_defintion()
and monitors_definition() and append the signal to the respectively list
required. Signal() now uses current_name and current_port to return the signal.
It was ensured that current_name and current_port are cleared before calling
a signal to avoid any issues.