Implement get_in() and get_out() in parser
Now the parser is fully funcitonal for files using single assignment
of inputs.

In get_out() is the only point we should find .IX to get the correct
port.id as the port is a number X we have to do a lookup in the names
module of the concatenattion IX. All other ports have the ids
corresponding to those assigned to the devices.

clear_vars() and clear_all() were implemented for more readability
and to avoid some bugs that occured using self.empty_symbol.

TO DO:
*Check for semantic errors in connections
*Implement one to many, many to one and many to many functionality