Update scanner to print errors on any line in the definition file
The scanner was able to print errors in the definition file using
get_error_line(), only if the symbol passed as an argument to get_error_line()
was at the same line as the current state of the scanner. In order to print
errors on any line, the scanner now keeps a record of the starting position of
all the lines it encounters in the definition file.