Fix bug of printing lines with tabs in get_error_line()
Lines that contain tabs get printed on the terminal with 8 spaces in the place
of the tab, by default. This was caussing misalignment of the printed line from
the circuit definition file and the pointer ("^").

This patch replaces any "\t" in the line to be printed with a single space, so
that the pointer and the line are now aligned.