Implemt one to many and one to many connections
The parser is able to parse correctly many to one connections.
Example: A, B, C => D; (where D is a NAND(3))
This is restricted only to gates as we do not care about the order of the
inputs whereas we care about the DTYPE. As well all the inputs
must be specificied at one.
Not allowed: A, B => D; C => D.I3;

It can also parse correctly one to many connections.
Example: A => B.I1, C.I1, D.I3;