Add update_zoom_lower_bound() method
This method updates the zoom lower bound such that the user cannot zoom out
further than the amount of zoom that enables all the signals to be seen at the
same time on the canvas.

For this to work smoothly, this method is called in the constructor, on a
resize event and in the render() method.