Fix bug related to monitors.get_margin() returning None
When the circuit definition file does not include any monitors, then the above
method returns None. This causes the render() method of the canvas to crash
because it is trying to update the canvas attribute margin_left with a
calcuation utilizing the method monitors.get_margin().

This patch fixes the issue.