Implement on_open() to actually open file & parse
When the open button is pressed and the file selected the instances
names, devices, devices, network, scanner and parser are initialised
to erase any previous data and it tries to parse the network.

GUI now only has one parameter title. The file to open is now selected
via the open button. logsim.py now does not take a file argument.