Fix bug related to open file button
When the open file button is pressed in the gui, and a new circuit definition
file is selected, then new instances of Names, Devices, Monitors, Scanner,
Parser are created in the gui instance. However, the canvas still has a reference
to the previous instances of the aforementioned classes.

This patch fixes the issue.