Fix errors on terminal on initialization of the canvas
The problem was that on opening the window for the first time, the on_size()
function of MyGLCanvas_2D was calling render() which in turn was trying to
access the openGL context before it was ready. This patch removes the call
to render() in the on_size() method and calls update_borders() and
bound_zooming() instead, to make the canvas look just right.